refactor: move slogdiscard to new location and update imports
- Move slogdiscard from internal/lib/logger/handler/slogdiscard to internal/lib/logger/slogdiscard
- Remove old logger structure (handler/slogpretty, logger.go)
- Update imports in test files to use new slogdiscard location
