--- internal/lib/logger/handler/slogpretty/slogpretty.go
@@ -1,103 +1,0 @@
-package slogpretty
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"io"
-	"log/slog"
-
-	"github.com/fatih/color"
-)
-
-type Options struct {
-	AddSource   bool
-	Level       slog.Leveler
-	ReplaceAttr func(groups []string, attr slog.Attr) slog.Attr
-}
-
-type PrettyHandler struct {
-	// opts Options
-	slog.Handler
-	attrs []slog.Attr
-}
-
-func NewPrettyHandler(
-	out io.Writer,
-	opt *Options,
-) *PrettyHandler {
-	h := &PrettyHandler{
-		Handler: slog.NewJSONHandler(out, &slog.HandlerOptions{
-			Level:     opt.Level,
-			AddSource: opt.AddSource,
-		}),
-	}
-
-	return h
-}
-
-func (h *PrettyHandler) Handle(_ context.Context, r slog.Record) error {
-	level := r.Level.String() + ":"
-
-	switch r.Level {
-	case slog.LevelDebug:
-		level = color.MagentaString(level)
-	case slog.LevelInfo:
-		level = color.BlueString(level)
-	case slog.LevelWarn:
-		level = color.YellowString(level)
-	case slog.LevelError:
-		level = color.RedString(level)
-	}
-
-	fields := make(map[string]interface{}, r.NumAttrs())
-
-	r.Attrs(func(a slog.Attr) bool {
-		fields[a.Key] = a.Value.Any()
-
-		return true
-	})
-
-	for _, a := range h.attrs {
-		fields[a.Key] = a.Value.Any()
-	}
-
-	var b []byte
-
-	var err error
-
-	if len(fields) > 0 {
-		b, err = json.MarshalIndent(fields, "", "  ")
-		if err != nil {
-			return err
-		}
-	}
-
-	timeStr := r.Time.Format("[15:05:05.000]")
-	msg := color.CyanString(r.Message)
-
-	fmt.Println(
-		timeStr,
-		level,
-		msg,
-		color.WhiteString(string(b)),
-	)
-
-	return nil
-}
-
-func (h *PrettyHandler) WithAttrs(attrs []slog.Attr) slog.Handler {
-	return &PrettyHandler{
-		Handler: h.Handler,
-		// l:       h.l,
-		attrs: attrs,
-	}
-}
-
-func (h *PrettyHandler) WithGroup(name string) slog.Handler {
-	// TODO: implement
-	return &PrettyHandler{
-		Handler: h.Handler.WithGroup(name),
-		// l:       h.l,
-	}
-}
